requirements.txt
openai
python-dotenv
requests
pydantic
watchdog
streamlit
streamlit-chat